Fix sign convertion for boringssl

+++ src/algorithm/crypto_dh.cpp
@@ -1237,7 +1237,7 @@ LIBATFRAME_UTILS_API int dh::shared_context::random(void *output, size_t output_
 
   int ret = error_code_t::OK;
 #  if defined(CRYPTO_USE_OPENSSL) || defined(CRYPTO_USE_LIBRESSL) || defined(CRYPTO_USE_BORINGSSL)
-  if (!RAND_bytes(reinterpret_cast<unsigned char *>(output), static_cast<int>(output_sz))) {
+  if (!RAND_bytes(reinterpret_cast<unsigned char *>(output), static_cast<size_t>(output_sz))) {
     ret = static_cast<int>(ERR_peek_error());
   }
